Fix 4th quadrant resistence bug by accounting for integer truncation in jitter domain [-1, 1] -> [-.5, 1.5].  Values between [0, 1] are "swallowed" during truncation. Better handle 0 sized allocs by always allocating 1 byte (technically not spec compliant). Entities now accure time increments (delta time) as opposed to dealing with absolute time. Introduce TimedScope to make easy measuring time duration in a scope. Minor code cleaup. Minor UBSAN related fixes.  Now runs clean w/ UBAN and ASAN.

+++ src/core/misc/random_number_distributions.hpp
@@ -12,8 +12,8 @@ SampleGaussian(RNGUni01& rng_ctx, f32 mean, f32 std_dev) {
 
 	f32 u, v, s;
 	do {
-		u = (f32)rng_ctx() * 2.0f - 1.0f;
-		v = (f32)rng_ctx() * 2.0f - 1.0f;
+		u = rng_ctx() * 2.0f - 1.0f;
+		v = rng_ctx() * 2.0f - 1.0f;
 		s = (u * u) + (v * v);
 	} while (s >= 1.0f or s == 0.0f);
 
@@ -36,8 +36,8 @@ SampleStandardGaussian(RNGUni01& rng_ctx) {
 
 	f32 u, v, s;
 	do {
-		u = (f32)rng_ctx() * 2.0f - 1.0f;
-		v = (f32)rng_ctx() * 2.0f - 1.0f;
+		u = rng_ctx() * 2.0f - 1.0f;
+		v = rng_ctx() * 2.0f - 1.0f;
 		s = (u * u) + (v * v);
 	} while (s >= 1.0f or s == 0.0f);
